requirements.txt
streamlit==1.46.1
tensorflow==2.19.0
# tensorflow-metal==1.2.0
# keras==3.9.2
numpy==2.1.3
pandas==2.2.3
pillow==11.2.1
plotly==6.2.0
scikit-learn==1.6.1
matplotlib==3.10.1